requirements.txt
discord.py==1.2.3
praw==6.4.0
python-dotenv==0.10.3
aiohttp==3.5.4
